break CapabilityTransactionalDecorator into CapabilityTransactionalDecorator, CapabilityOutboxEnabledDecorator, TopicTransactionalDecorator and TransactionalDecoratorBase
